Refactor opening data structures
