+++ crates/review-domain/src/opening/mod.rs
@@ -1,0 +1,9 @@
+//! Opening-specific data structures.
+
+mod card;
+mod edge;
+mod edge_input;
+
+pub use card::OpeningCard;
+pub use edge::OpeningEdge;
+pub use edge_input::EdgeInput;
